% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRD_utility.R
\name{utilsCalculateRank}
\alias{utilsCalculateRank}
\title{utilsCalculateRank}
\usage{
utilsCalculateRank(df, nameCol)
}
\arguments{
\item{df}{A DataFrame.}

\item{nameCol}{The name of the column to be ranked. Note that this 
parameter needs to be specified as there is no default value.}
}
\value{
Returns a new \code{df} that has an additional column with
the rankings of the column specified by \code{nameCol}.
}
\description{
Calculates the ranking of a given column.
}
\examples{
SRDInput <- data.frame(
A=c(32, 52, 44, 44, 47),
B=c(73, 75, 65, 76, 70),
C=c(60, 59, 57, 55, 60),
D=c(35, 24, 44, 83, 47),
E=c(41, 52, 46, 50, 65))
columnName <- "A"
rSRD::utilsCalculateRank(SRDInput,columnName)
}
\author{
Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}
}
