% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRD_utility.R
\docType{data}
\name{utilsColorPalette}
\alias{utilsColorPalette}
\title{utilsColorPalette}
\format{
An object of class \code{character} of length 250.
}
\usage{
utilsColorPalette
}
\description{
Unique color palette for heatmaps.
}
\examples{
barplot(rep(1,250), col = utilsColorPalette)
}
\author{
Attila Gere \email{gereattilaphd@uni-mate.hu}, Balázs R. Sziklai \email{sziklai.balazs@krtk.hu},
}
\keyword{datasets}
