% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{utilsMaxSRD}
\alias{utilsMaxSRD}
\title{utilsMaxSRD}
\usage{
utilsMaxSRD(rowsCount)
}
\arguments{
\item{rowsCount}{The number of rows in the SRD calculation.}
}
\value{
The maximum achievable SRD value.
}
\description{
Calculates the maximum distance between two rankings of size n.
This function is used to normalize SRD values.
}
\examples{
maxSRD <- rSRD::utilsMaxSRD(5)
}
\author{
Dennis Horn
}
