% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fusion.R
\name{utilsPreprocessDF}
\alias{utilsPreprocessDF}
\title{utilsPreprocessDF}
\usage{
utilsPreprocessDF(df, method = "range_scale")
}
\arguments{
\item{df}{A DataFrame.}

\item{method}{A string that should contain "scale_to_unit", "standardize", "range_scale" or "scale_to_max".}
}
\value{
Returns a new \code{df} that has a Distance Column based on the \code{nameCol}.
}
\description{
This function preprocesses the DataFrame depending on the \code{method}.
}
\examples{
SRDInput <- data.frame(
A=c(32, 52, 44, 44, 47),
B=c(73, 75, 65, 76, 70),
C=c(60, 59, 57, 55, 60),
D=c(35, 24, 44, 83, 47),
E=c(41, 52, 46, 50, 65))
method <- "standardize"
utilsPreprocessDF(SRDInput,method)
}
\author{
Ali Tugay Sen, Dennis Horn \email{dennishorn@hotmail.de}, Linus Olsson \email{linusmeol@gmail.com}
}
