% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utilsRankingMatrix}
\alias{utilsRankingMatrix}
\title{utilsRankingMatrix}
\usage{
utilsRankingMatrix(data_matrix)
}
\arguments{
\item{data_matrix}{A DataFrame.}
}
\value{
A DataFrame containing the ranking matrix.
}
\description{
R interface to perform the rank transformation on the columns of the input data frame. 
Ties are resolved by fractional ranking.
}
\examples{
df <- data.frame(
A=c(32, 52, 44, 44, 47),
B=c(73, 75, 65, 76, 70),
C=c(60, 59, 57, 55, 60),
D=c(35, 24, 44, 83, 47),
E=c(41, 52, 46, 50, 65))

utilsRankingMatrix(df)
}
\author{
Balázs R. Sziklai \email{sziklai.balazs@krtk.hu}, Linus Olsson \email{linusmeol@gmail.com}
}
