% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRD_utility.R
\name{utilsTieProbability}
\alias{utilsTieProbability}
\title{utilsTieProbability}
\usage{
utilsTieProbability(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Returns the tie probability as a numeric value.
}
\description{
Calculates the tie probability for a given vector. The tie probability is defined as the number of consecutive tied component-pairs \emph{in the sorted vector} divided by the size of the vector minus 1.
}
\examples{
x <-c(1,2,4,4,5,5,6)
rSRD::utilsTieProbability(x)
}
\author{
Ali Tugay Sen, Linus Olsson \email{linusmeol@gmail.com}
}
